Added the 'stage' system
NPC's teleport to the respective stage flag, enter a FIFO queue for processing, when processing starts, timer counts down and then generates a random float from 0-1, if fail the chance for success is incremented and the timer resets. iterated until success and npc moves to next stage. this is done until NPC's reach the doneALL state which represents boarded/at the gate